add delect service and validate the input

+++ event/src/services/event.service.ts
@@ -1,5 +1,7 @@
 import { EventDetail } from "@user/databases/@types/event.interface";
 import { EventRepository } from "@user/databases/repositories/event.reposities";
+import APIError from "@user/Errors/api-error";
+import { StatusCode } from "@user/utils/consts";
 
 export class EventService {
   public eventRepo: EventRepository;
@@ -17,6 +19,12 @@ export class EventService {
 
   async updateEvent(id: string, eventDetail: EventDetail) {
     try {
+      const existedEvent = await this.eventRepo.findEvent(id);
+
+      if (!existedEvent) {
+        throw new APIError("Event not found", StatusCode.NotFound);
+      }
+
       return await this.eventRepo.updateEvent(id, eventDetail);
     } catch (error: unknown) {
       throw error;
@@ -25,7 +33,21 @@ export class EventService {
 
   async deleteEvent(id: string) {
     try {
+      const existedEvent = await this.eventRepo.findEvent(id);
+
+      if (!existedEvent) {
+        throw new APIError("Event not found", StatusCode.NotFound);
+      }
+
       return await this.eventRepo.deleteEvent(id);
+    } catch (error: unknown) {
+      throw error;
+    }
+  }
+
+  async findEvent(id: string) {
+    try {
+      return await this.eventRepo.findEvent(id);
     } catch (error: unknown) {
       throw error;
     }
